+++ main.c
@@ -1,10 +1,24 @@
 
-#include <winstl/diagnostics/output_debug_line.h>
+#include <stlsoft/stlsoft.h>
+
+#if _STLSOFT_VER < 0x010b0184
+# error requires STLSoft v1.11.1-b4 or later
+#endif
+
 #include <platformstl/filesystem/path_functions.h>
+#include <winstl/diagnostics/output_debug_line.h>
+
+#include <windows.h>
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+
+
+#define WRITEDEBUGSTRING_VER_MAJOR  0
+#define WRITEDEBUGSTRING_VER_MINOR  0
+#define WRITEDEBUGSTRING_VER_PATCH  1
+
 
 int main(int argc, char* argv[])
 {
@@ -27,8 +41,22 @@ missing_arguments:
     {
         fprintf(
             stdout
-        ,   "USAGE: %.*s { --help | [ --sleep-after-writes ] <msg-1> [ ... <msg-N> ] }\n"
+        ,   "USAGE: %.*s { --help | --version | [ --sleep-after-writes ] <msg-1> [ ... <msg-N> ] }\n"
         ,   (int)program_name.len, program_name.ptr
+        );
+
+        return EXIT_SUCCESS;
+    }
+    else
+    if (0 == strcmp("--version", argv[1]))
+    {
+        fprintf(
+            stdout
+        ,   "%.*s %d.%d.%d\n"
+        ,   (int)program_name.len, program_name.ptr
+        ,   WRITEDEBUGSTRING_VER_MAJOR
+        ,   WRITEDEBUGSTRING_VER_MINOR
+        ,   WRITEDEBUGSTRING_VER_PATCH
         );
 
         return EXIT_SUCCESS;
@@ -42,7 +70,7 @@ missing_arguments:
         {
             char const* const arg = argv[i];
 
-            if (0 == strcmp("--sleep-after-writes", argv[i]))
+            if (0 == strcmp("--sleep-after-writes", arg))
             {
                 do_sleep = 1;
             }
@@ -69,4 +97,3 @@ missing_arguments:
         return EXIT_SUCCESS;
     }
 }
-